Use BigDecimal for statistics model classes

+++ reports-reloaded/src/main/java/com/novoda/github/reports/data/mock/MockProjectDataLayer.java
@@ -3,11 +3,20 @@ package com.novoda.github.reports.data.mock;
 import com.novoda.github.reports.data.ProjectDataLayer;
 import com.novoda.github.reports.data.model.ProjectRepoStats;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class MockProjectDataLayer implements ProjectDataLayer {
     @Override
     public ProjectRepoStats getStats(String project, Date from, Date to) {
-        return new ProjectRepoStats(project, 12, 13, 48, 96, 123, 8);
+        return new ProjectRepoStats(
+                project,
+                BigDecimal.valueOf(12),
+                BigDecimal.valueOf(13),
+                BigDecimal.valueOf(48),
+                BigDecimal.valueOf(96),
+                BigDecimal.valueOf(123),
+                BigDecimal.valueOf(8)
+        );
     }
 }
